(SERVER) Add post authorizations

+++ server/src/modules/post/post.service.ts
@@ -1,14 +1,15 @@
-import Post, { IPost } from '../../models/post';
 import cuid from 'cuid';
 import slug from 'limax';
 import sanitizeHtml from 'sanitize-html';
+import Post, { IPost } from '../../models/post';
+import CustomError from '../../utils/errors/CustomError';
 
 export const getPosts = async (): Promise<IPost[]> => {
     return Post.find().sort('-dateAdded');
 }
 
-export const addPost = async (post: { name: string, title: string, content: string }): Promise<IPost> => {
-    const newPost: IPost = new Post(post);
+export const addPost = async (post: { name: string, title: string, content: string }, userId: string): Promise<IPost> => {
+    const newPost: IPost = new Post({ ...post, user_id: userId });
     newPost.title = sanitizeHtml(newPost.title);
     newPost.name = sanitizeHtml(newPost.name);
     newPost.content = sanitizeHtml(newPost.content);
@@ -22,6 +23,14 @@ export const getPost = async (cuid: string): Promise<IPost | null> => {
     return Post.findOne({ cuid });
 }
 
-export const deletePost = async (cuid: string): Promise<{ deletedCount: number }> => {
+export const deletePost = async (cuid: string, userId: string): Promise<{ deletedCount: number }> => {
+    const post = await Post.findOne({ cuid });
+    if (!post) {
+        return { deletedCount: 0 };
+    }
+
+    if (post.user_id !== userId) {
+        throw new CustomError('User cant delete other user posts', 403);
+    }
     return Post.deleteOne({ cuid });
 }
